Relocated Spec Catalogs to be made available outside of gobblin-service
